feature(async): move useRouter to separate file
